Rename 'vectordb' folder to 'vector_db'
